🚚 Rename dialogs.ts -> ui.dialogs.ts to group w/ notifs
