+++ test/unit/mods.example.multitype.ts
@@ -2,7 +2,7 @@
 // to match things, and normally within them.
 
 import path from "path";
-import { InstallChoices } from "../../src/dialogs";
+import { InstallChoices } from "../../src/ui.dialogs";
 import {
   CONFIG_XML_MOD_BASEDIR,
   CONFIG_JSON_MOD_BASEDIR_SETTINGS,
